добавил wrapper'ы для фильтра данных

+++ app/src/main/java/ru/modelov/testapp65apps/main/entities/Specialty.kt
@@ -10,7 +10,7 @@ typealias Specialties = List<Specialty>
 data class Specialty(
     @PrimaryKey
     @SerializedName("specialty_id")
-    val specialtyId: Int,
+    val specialtyId: Long,
 
     @SerializedName("name")
     val name: String
